requirements-all.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.35
asyncpg==0.30.0
pydantic==2.9.0
pydantic-settings==2.5.0
google-generativeai
python-jose[cryptography]==3.3.0
passlib[bcrypt]
httpx==0.27.0
python-multipart==0.0.12
GeoAlchemy2==0.15.0
pytest==8.3.0
pytest-asyncio==0.24.0
pytest-cov==5.0.0
alembic==1.13.2
psycopg2-binary==2.9.9
shapely==2.0.5
numpy==2.0.2
